+++ RELEASE-NOTES.md
@@ -12,7 +12,7 @@
   * Added `image-registry login` command to fetch authentication token and log in to image registry in one command.
   * Added `image-repository url <repo_name>` command to get the URL for specified image repository.
   * Added `create` command for `image-repository`.
-  * Added `set (property)`, `unset (property)`, `suspend` and `resume` commands for `compute-pool`.
+  * Added `status`, `set (property)`, `unset (property)`, `suspend` and `resume` commands for `compute-pool`.
   * Added `set (property)`, `unset (property)`,`upgrade` and `list-endpoints` commands for `service`.
 * Connections parameters are also supported by generic environment variables:
   * `SNOWFLAKE_ACCOUNT`